moved background worker into own file
