+++ src/main.rs
@@ -1,3 +1,4 @@
+mod background_worker;
 mod change_state;
 mod check_email;
 mod check_pdf;
@@ -15,6 +16,7 @@ mod templates;
 mod view_status;
 mod view_status_admin;
 
+use crate::background_worker::background_worker;
 use crate::change_state::change_state;
 use crate::close_popup::close_modal;
 use crate::home::home;
@@ -29,54 +31,9 @@ use actix_files as fs;
 use actix_multipart::form::tempfile::TempFileConfig;
 use actix_web::web::PayloadConfig;
 use actix_web::{App, HttpServer, web};
-use chrono::Local;
+use sqlx::Sqlite;
 use sqlx::migrate::MigrateDatabase;
 use sqlx::sqlite::SqlitePoolOptions;
-use sqlx::{Row, Sqlite, SqlitePool};
-use tokio::time::{Duration, interval};
-
-/// Runs every 24 hours on a background thread
-pub async fn background_worker(pool: SqlitePool) {
-    let mut ticker = interval(Duration::from_hours(24));
-
-    loop {
-        ticker.tick().await;
-        println!("Running scheduled task...");
-
-        // Get current time as UTC
-        let current_time = Local::now().naive_local();
-
-        // Get the rows
-        let rows = sqlx::query("SELECT key, value, state, created_at FROM applicants")
-            .fetch_all(&pool)
-            .await
-            .unwrap();
-
-        for row in rows {
-            let key: String = row.get("key");
-            let created_at: String = row.get("created_at");
-            let state: i32 = row.get("state");
-            let user_created_at: chrono::NaiveDateTime =
-                chrono::NaiveDateTime::parse_from_str(created_at.as_str(), "%Y-%m-%d %H:%M:%S")
-                    .unwrap();
-
-            let duration_in_db = current_time - user_created_at;
-
-            // If the duration equals one day and the state is 'fresh' it will be set to 'old'
-            if duration_in_db.num_days() == 1 && state == 1 {
-                let new_state = state + 1;
-
-                // Update the state in the db
-                sqlx::query("UPDATE applicants SET state = ? WHERE key = ?")
-                    .bind(new_state)
-                    .bind(key)
-                    .execute(&pool)
-                    .await
-                    .unwrap();
-            }
-        }
-    }
-}
 
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
